+++ src/main/java/org/openas2/processor/storage/ArchiveStorageModule.java
@@ -1,0 +1,13 @@
+package org.openas2.processor.storage;
+
+public class ArchiveStorageModule extends MessageFileModule
+{
+	/**
+	 * {@inheritDoc}
+	 */
+	@Override
+	protected String getModuleAction()
+	{
+		return DO_ARCHIVE;
+	}
+}