+++ src/main/java/org/openas2/processor/storage/StorageModule.java
@@ -6,5 +6,7 @@ public interface StorageModule extends ProcessorModule
 {
 	public static final String DO_STORE = "store";
 
-	public static final String DO_STOREMDN = "storemdn";
+	public static final String DO_STOREMDN = "store_mdn";
+
+	public static final String DO_ARCHIVE = "store_archive";
 }